hocr-pagenumbers: fill up missing numbers with numbers from phase 1
In some cases the formal model rejects valid page numbers (for example
if ~10 pages of a ~300 page book are suddenly in a completely different
place). This is hard to fix, since the features do make them outliers.
We can rely on the results from the first phase to at least fill in the
numbers, even though they are less accurate.