hocr-pagenumbers: allow composite page numbers to be larger than 10000
The base for the composite numbers was picked according to the paper,
which describes 10000 as "sufficiently high", but this doesn't seem to
be case for us, as there are single-digiti composite page number
*candidates* that are larger than 10000 - which then make the code think
there are actually two digits because the number exceeds the base.

For now, fix this by using the fact that Python allows arbitrarily large
integers, so use 1000000000000 as base for now.