hocr-pagenumbers: raise arbitrary composite scheme limit
We were semi regularly hitting the limit of composite schemes on books
that have many pages and also have a lot of text per page. Maybe we can
eventually discard scheme's that aren't seen for more than a certain
amount of pages (and also have just one match), but for now let's just
raise the limit.