hocr-to-epub: ignore all covers if one is cloth
commit 6912b4ed157821aa428ff3424fc71deea2836f5e ("hocr-to-epub: ignore
cloth covers") introduced a mechanism to ignore certain covers if the
cover is detected as cloth cover. This commit changes the logic to
discard all cover pages if one of the pages is detected as cloth cover,
as it is likely that the other pages marked as cloth cover are also
cloth covers (but might not be detected as such) - such as the back
cover of a book.