Fix OPF-043
`epubcheck` `v5.1.0`, validating `EPUB 3.3`, took issue with the
`text/html` `media_type` for the spine.

This commit fixes that.

Fixing this error created another one, insofar as HTML within the next
begun to be rendered, and that caused another problem. E.g., if the OCR
recognized text as saying, `<A`, the epub generator would try to create
a link.

See, e.g., the OCRed text, `<A bicker of musketry fire` within
https://archive.org/download/sim_english-illustrated-magazine_1884-12_2_15/sim_english-illustrated-magazine_1884-12_2_15_hocr.html

The fix for this was to HTML escape text.